# ASV: used integer ndarray as indexer for Series.take indexing benchmark