@hardikpnsp 2020-09-26
#### Updated benchmarks after changing indexer from list to ndarray:

```
[100.00%] ··· indexing.Take.time_take                                                                                                                                                ok
[100.00%] ··· ========== =========
                index             
              ---------- ---------
                 int      336±2μs 
               datetime   348±2μs 
              ========== =========
```